% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apple.R
\docType{data}
\name{apple}
\alias{apple}
\title{HRATA data on apple}
\format{
a dataset with X wines, 2 replicates and 33 tasters
\describe{
\item{CJ}{33 tasters from ESA Angers - MIM ASPE 2018}
\item{ProductName}{6 wines}
\item{Attributes}{23 sensory attibutes scored from 0 to 10}
}
}
\usage{
data(apple)
}
\description{
Data from a tasting done on 12 apples with X panelis from ESA Angers in 2020
}
\keyword{datasets}
