% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonisation.R
\name{canonisation}
\alias{canonisation}
\title{canonistation : transpose a multivariable table in one column}
\usage{
canonisation(data.canon, NbLeft)
}
\arguments{
\item{data.canon}{a dataframe to be change from a multiple column to a two column wit a replication of the first ones}

\item{NbLeft}{a value indicating the number of column on the left of the table that do no be concatened}
}
\value{
XXXXXX
}
\description{
transpose a multivariable table in one column
}
