% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Descri.Quanti.R
\name{Descri.Quanti}
\alias{Descri.Quanti}
\title{Descri.Quanti : Boxplot  and graph for numeric variables}
\usage{
Descri.Quanti(
  x,
  column,
  group = NULL,
  print.graph = TRUE,
  graph.grouped = TRUE,
  save.graph = FALSE,
  date = FALSE,
  extension = "wmf"
)
}
\arguments{
\item{x}{dataframe}

\item{column}{column to be describe}

\item{group}{name of the column to use for grouping data}

\item{print.graph}{logical. plot the graphs. TRUE by default}

\item{graph.grouped}{logical. group th graph into one. TRUE by default}

\item{save.graph}{logical. save the graphe. TRUE by default}

\item{date}{logical. include date in the name of file : FALSE by default}

\item{extension}{extension for the images(wmf, png, etc.)}
}
\description{
Descri.Quanti : Boxplot  and graph for numeric variables
}
\examples{
data(wine)
Descri.Quanti(wine, c(3:5), "ProductName")

}
