% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.AOV.R
\name{export.AOV}
\alias{export.AOV}
\title{export.AOV : Export results from AOV.Grappe}
\usage{
export.AOV(
  res.from.AOV,
  save.stat = TRUE,
  save.LSD = TRUE,
  name = substitute(res.from.AOV),
  name.stat = NULL,
  name.LSD = NULL,
  print.dir = TRUE
)
}
\arguments{
\item{res.from.AOV}{results from \code{\link[=AOV.Grappe]{AOV.Grappe()}}}

\item{save.stat}{save pvalue, signif, F, DF et models results. TRUE by default}

\item{save.LSD}{save LS means results. TRUE by defaultTRUE by default}

\item{name}{name of the existing files}

\item{name.stat}{name for the Stat Files (defaut is NULL)}

\item{name.LSD}{name for the LSD Files (defaut is NULL)}

\item{print.dir}{print the directory name. TRUE by default}
}
\description{
Export results from \code{\link[=AOV.Grappe]{AOV.Grappe()}}
}
\examples{
data(wine)
Var.Grappe(wine)
winef=Var.Grappe(wine,column = c(1:2), type="factor")
res.AOV <- AOV.Grappe( x=winef,column = c(3:4), "ProductName + (1|CJ)+ (1|ProductName:CJ)")
export.AOV(res.AOV)
library(RColorBrewer)
mycolor=brewer.pal(7,"Greens")
graph.AOV(res.AOV,color.graph=mycolor)
graph.AOV(res.AOV,title.graph="Sensory Score", add.moy=TRUE, x.ordered=1, ymin=1, ymax=9)


}
\seealso{
\code{\link[=AOV.Grappe]{AOV.Grappe()}}
}
