% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcata.table.R
\name{hcata.table}
\alias{hcata.table}
\title{hcata.table : prepare the hcata table for analysis}
\usage{
hcata.table(res.agreg)
}
\arguments{
\item{res.agreg}{a dataframe xxxxxxxx}
}
\value{
returns graphs of means by factors and variables with letters from posthoc test, standard deviations and pvalue from ANOVAs
}
\description{
create the dataframe use for HCATA analysis
}
\examples{
data(apple)
data(apple.attribute)
res.agreg<-hrata.agregation(data=apple,h.table=apple.attribute,crit.agreg=max)
res.signi<-hrata.signi(res.agreg,seuil=0.05)
res.table<-hcata.table(res.agreg)
res.PCA<-hrata.multidim(res.table,method="CA",niv=3)

}
\seealso{
xxxxxx
}
