% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrata.agregation.R
\name{hrata.agregation}
\alias{hrata.agregation}
\title{hrata.agregation : prepare the hrata table for analysis}
\usage{
hrata.agregation(data, h.table, crit.agreg = max)
}
\arguments{
\item{data}{a dataframe containing rawdata HRATA or HCATA sensory task wiht Juge in first column and Product in the second column}

\item{h.table}{a dataframe with the hierarchical structure : three columns Familly, Category, Terms}

\item{crit.agreg}{agregation criterion used during the data allocation : max, min, or mean}
}
\value{
returns graphs of means by factors and variables with letters from posthoc test, standard deviations and pvalue from ANOVAs
}
\description{
create the dataframe use for HRATA analysis
}
\examples{
data(apple)
data(apple.attribute)
res.agreg<-hrata.agregation(data=apple,h.table=apple.attribute,crit.agreg=max)
res.signi<-hrata.signi(res.agreg,seuil=0.05)
res.table<-hrata.table(res.agreg,type="dravnieks")
res.PCA<-hrata.multidim(res.table,method="PCA",scale.unit=FALSE,niv=3)


}
\seealso{
xxxxxx
}
