% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrata.codage.R
\name{hrata.codage}
\alias{hrata.codage}
\title{hrata.codage : function to modify Hierachical structure CSV file}
\usage{
hrata.codage(h.table)
}
\arguments{
\item{h.table}{dataframe from a CSV file containing the Hierachical structure}
}
\description{
hrata.codage : function to modify Hierachical structure CSV file
}
\examples{
data(apple.attribute)
hrata.codage(apple.attribute)


}
