% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrata.signi.R
\name{hrata.signi}
\alias{hrata.signi}
\title{hrata.signi : Perform a regression analysis}
\usage{
hrata.signi(res.agreg, seuil = NULL)
}
\arguments{
\item{res.agreg}{results from \link{hrata.agregation}}

\item{seuil}{the significance threshold considered for the analyses of variance. Only attribute under this value will keep in the results (by default 0.05)}
}
\value{
Returns graphs of means by factors and variables with letters from posthoc test, standard deviations and pvalue from ANOVAs
}
\description{
Perform a XXXXX analysis on the HRATA or HCATA data to identify discriminative attibutes
}
\examples{
data(apple)
data(apple.attribute)
res.agreg<-hrata.agregation(data=apple,h.table=apple.attribute,crit.agreg=max)
res.signi<-hrata.signi(res.agreg,seuil=0.05)
res.table<-hrata.table(res.agreg,type="dravnieks")
res.PCA<-hrata.multidim(res.table,method="PCA",scale.unit=FALSE,niv=3)


}
\seealso{
\code{\link[=AOV.Grappe]{AOV.Grappe()}}
}
