% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose.attribute.R
\docType{data}
\name{rose.attribute}
\alias{rose.attribute}
\title{Hierarchical structure of the attributes in HRATA rose dataset}
\format{
a dataframe with 3 columns (Family,Category and Attribute)
\describe{
\item{Family}{Aromatic families}
\item{Category}{Categories under the families}
\item{Attribute}{Attributes or terms under category and family}
}
}
\usage{
data(rose.attribute)
}
\description{
Dataframe of the rose flower aromatic wheel structured with three columns and presenting the hierarchical organization of the dataset (wheel)
Family and Category are duplicated to express the hierarchical structure
}
\keyword{datasets}
