% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose.R
\docType{data}
\name{rose}
\alias{rose}
\title{HRATA data on rose flower}
\format{
a dataset with 9 rose bushes and 74 tasters
\describe{
\item{Juge}{74 tasters recruited by ESA Angers - UR GRAPPE}
\item{Product}{9 rose cultivars}
\item{Attributes}{118 sensory attributes scored from NA to 3 following the HRATA methodology}
}
}
\usage{
data(rose)
}
\description{
Data from a tasting done on 9 rose cultivars with 74 tasters in 2023
}
\keyword{datasets}
