% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.Grappe.R
\name{STAR.Grappe}
\alias{STAR.Grappe}
\title{STAR.Grappe : function for encode pvalue to stars}
\usage{
STAR.Grappe(values)
}
\arguments{
\item{values}{dataframe, vector or column of a matrix with pvalue to be coded into stars}
}
\description{
STAR.Grappe : function for encode pvalue to stars
}
\examples{
attach(iris)
STAR.Grappe(iris$Sepal.Width)

}
