% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.get.binary.R
\name{transfert.MV.BM}
\alias{transfert.MV.BM}
\title{Décomposition d'un descripteur hiérarchique (format ConsoTextPlorer)}
\usage{
transfert.MV.BM(x, tres = FALSE)
}
\arguments{
\item{x}{Caractère. Chaîne à décomposer, au format
\code{"dimension:concept/descripteur"}. Peut être \code{NA}.}

\item{tres}{Logique. Si \code{FALSE} (défaut), le suffixe \verb{_très} est
supprimé des composantes \code{concept} et \code{descripteur}.
Si \code{TRUE}, il est conservé.}
}
\value{
Une matrice de dimensions 1 × 3 contenant, dans l'ordre :
\code{dimension}, \code{concept}, \code{descripteur}. Si \code{x} est \code{NA}, retourne
la matrice \code{c("rien", "rien", "rien")}.
}
\description{
Décompose une chaîne de caractères au format ConsoTextPlorer
(\code{"dimension:concept/descripteur"}) en ses trois composantes :
dimension, concept et descripteur.
}
\details{
Le format attendu est \code{"dimension:concept/descripteur"}, où :
\itemize{
\item \code{":"} sépare la dimension du concept ;
\item \code{"/"} sépare le concept du descripteur.
}
Si le descripteur vaut \code{"_"}, il est remplacé par la valeur du concept.
}
\examples{
transfert.MV.BM("texture:croustillant/très_croustillant", tres = FALSE)
#>      [,1]      [,2]           [,3]
#> [1,] "texture" "croustillant" "croustillant"

transfert.MV.BM("texture:croustillant/très_croustillant", tres = TRUE)
#>      [,1]      [,2]           [,3]
#> [1,] "texture" "croustillant" "très_croustillant"

transfert.MV.BM(NA)
#>      [,1]    [,2]    [,3]
#> [1,] "rien"  "rien"  "rien"

}
\seealso{
\code{\link{get.binary}}, \code{\link{data.preprocess.fc}}
}
