% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Var.Grappe.R
\name{Var.Grappe}
\alias{Var.Grappe}
\title{Var.Grappe : Checking and encoding dataset}
\usage{
Var.Grappe(x, column = 0, type = NULL, filter = NULL, verbose = TRUE)
}
\arguments{
\item{x}{dataframe}

\item{column}{the numbers of the columns to change the type}

\item{type}{expected type : "factor", "numeric" or "integer"}

\item{filter}{text for searching name in column (letter or number)}

\item{verbose}{Display of the results ? TRUE by default}
}
\description{
Check dataset and  encode easily the vector (variables) to factor, integer or others
}
\examples{
data(wine)
Var.Grappe(wine)
Var.Grappe(wine, filter="O_")
winef=Var.Grappe(wine,column = c(1:2), type="factor")

}
