% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine.R
\docType{data}
\name{wine}
\alias{wine}
\title{Sensory evaluation of wine}
\format{
a dataset with 6 wines, 2 replicates and 33 tasters
\describe{
\item{CJ}{33 tasters from ESA Angers - MIM ASPE 2018}
\item{ProductName}{6 wines}
\item{Attributes}{23 sensory attibutes scored from 0 to 10}
}
}
\usage{
data(wine)
}
\description{
Data from a tasting done with X students from ESA Angers in 2015
}
\keyword{datasets}
