% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOV.Grappe.R
\name{AOV.Grappe}
\alias{AOV.Grappe}
\title{AOV.Grappe  ANOVA et post-hoc analysis for Linear Mixed Models}
\usage{
AOV.Grappe(x,
 column,
  Model,
  Random = TRUE,
  randomodel=F,
  reducF=FALSE,
  reducR=FALSE,
  graphic=FALSE,
  verbose=TRUE)
}
\arguments{
\item{x}{dataframe}

\item{column}{number of the columns on which ANOVA are performed}

\item{Model}{anova Model
- Syntax for fixed factor                :    facteur_A
- Syntax for Fixed factorsr            : 1|facteur_A)
- Fixed factors                          : facteur1 + facteur2
- Interaction for Fixed Factors    : facteur1 + facteur2 + facteur1:facteur2
- Fixed factor & Fixed factor      : facteur1 + (1|facteur2)
- Interaction with random          : (1|facteur1:facteur2)"}

\item{Random}{Not use for the moment (to be check)}

\item{randomodel}{Not use for the moment (to be check)}

\item{reducF}{logical.  reduce fixed effect structure? FALSE by default.}

\item{reducR}{logical. reduce random effect structure? FALSE by default.}

\item{graphic}{plot the graph TRUE / FALSE}

\item{verbose}{logical.show the progression of the analysis ? TRUE by Default}
}
\description{
ANOVA table with F-tests and p-values using Satterthwaite's method for denominator
degrees-of-freedom and F-statistic. Models should be fitted with \code{\link[=lmer]{lmer()}} from the lmerTest-package.
Then  LS-means differences of LS-mean for all considered factors in the linear mixed model.
}
\examples{
data(wine)
Var.Grappe(wine)
winef=Var.Grappe(wine,column = c(1:2), type="factor")
res.AOV <- AOV.Grappe( x=winef,column = c(3:4), "ProductName + (1|CJ)+ (1|ProductName:CJ)")
export.AOV(res.AOV)
library(RColorBrewer)
mycolor=brewer.pal(7,"Greens")
graph.AOV(res.AOV,color.graph=mycolor)
graph.AOV(res.AOV,title.graph="Sensory Score", add.moy=TRUE, x.ordered=1, ymin=1, ymax=9)



}
