% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apple.attribute.R
\docType{data}
\name{apple.attribute}
\alias{apple.attribute}
\title{Hierarchical structure}
\format{
a XXXX
\describe{
\item{CJ}{33 tasters from ESA Angers - MIM ASPE 2018}
\item{ProductName}{6 wines}
\item{Attributes}{23 sensory attibutes scored from 0 to 10}
}
}
\usage{
data(apple.attribute)
}
\description{
xxxxxx
}
\keyword{datasets}
