% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonisation.R
\name{canonisation}
\alias{canonisation}
\title{canonisation : transpose a multivariable table in one column}
\usage{
canonisation(data.canon, NbLeft)
}
\arguments{
\item{data.canon}{a dataframe to be change from a multiple column to a two column wit a replication of the first ones}

\item{NbLeft}{a value indicating the number of column on the left of the table that do no be concatened}
}
\value{
a dataframe with the NbLeft columns, a column with the canonized variables in "Attribute" column and the value in "Score" column
}
\description{
transpose a multivariable table in one column
}
\examples{
data(wine)
Var.Grappe(wine)
winef <- Var.Grappe(wine, column = c(1:2), type = "factor")
canonisation(winef, 2)

}
