% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.AOV.R
\name{graph.AOV}
\alias{graph.AOV}
\title{graph.AOV : Plot Means associated to AOV.Grappe function}
\usage{
graph.AOV(
  AOV.res = NULL,
  factor = NULL,
  attribute = NULL,
  horizontal = FALSE,
  pvalue = TRUE,
  title.graph = NULL,
  ymin = 0,
  ymax = NULL,
  angle.x = 0,
  angle.y = 0,
  angle.grp = 0,
  size.x = 10,
  size.grp = 4,
  save.graph = FALSE,
  x.ordered = 3,
  prefix = NULL,
  suffix = NULL,
  add.moy = FALSE,
  extension = "wmf",
  color.graph = NULL,
  y.label = "Mean",
  x.label = NULL
)
}
\arguments{
\item{AOV.res}{an object results from \code{\link[=AOV.Grappe]{AOV.Grappe()}}}

\item{factor}{factor is used to precise the fixed factor(s) to be presented. NULL by default
- factor=NULL then all factors are presented
- to represent one factor use  factor="ProductName"
- to represent several factos use   factor=c("ProductName", "Block", "Seance")}

\item{attribute}{attribute is used to precise the attribute(s) to be presented. NULL by default
- Attribute=NULL then all attributes are presented
- to represent one attribute use  attribute="S_sucre"
- to represent several attributes use   attribute=c("S_sucre", "A_abricot")}

\item{horizontal}{logical.draw horizontal graphs ( FALSE by default)}

\item{pvalue}{logical.add the pvalue of the representated factor on the graph ? TRUE by default}

\item{title.graph}{Value to be given for all title of each graph. By default, title.graph = NULL makes each title presenting the attribute name}

\item{ymin}{minimum for Y (0 by default)}

\item{ymax}{maximum  for Y (by default the max of Y is used)}

\item{angle.x}{angle for labels on X (by default 0) but 45 is better for long labels
- if horizontal=TRUE, then labels are oriented automatically}

\item{angle.y}{angle for labels on y (by default 0)}

\item{angle.grp}{angle for the LSD test letters over the histogram (by default 0)}

\item{size.x}{size of the fonts for abscisse}

\item{size.grp}{size of the fonts for the group letters}

\item{save.graph}{logical.Save the graph in the working directory  ? FALSE by default
-to access the working directory you can use : getwd()}

\item{x.ordered}{ordered abscisse axis (by default = 3)}

\item{prefix}{prefix for the name of the saved files}

\item{suffix}{suffix for the name of the saved files}

\item{add.moy}{logical.  add the value of the mean on the graph (FALSE by default)}

\item{extension}{image format (wmf, png, etc.) wmf by default}

\item{color.graph}{vector containing as much color than needed (NULL by defaut)}

\item{y.label}{value for the label (by default : "Mean")}

\item{x.label}{value for the label (by default : "Product")}
}
\value{
Returns graphs of means by factors and variables with letters from posthoc test, standard deviations and pvalue from ANOVAs
}
\description{
Plotting method for results obtain with AOV.Grappe. Mean value are represented associated with post-hoc test results done using lmerTest. Fixed Factors and Interactions can be drawn
}
\details{
By default, the function uses the 20 colors from the command = c(brewer.pal(12,"Set3"),brewer.pal(8,"Set1")) that is recorded in the function.
It is possible to modify by entering a color vector containing at least the same number of colors to use in the graph
For example :  color.graph= rep("grey",6)
}
\examples{
data(wine)
Var.Grappe(wine)
winef <- Var.Grappe(wine, column = c(1:2), type = "factor")
res.AOV <- AOV.Grappe(x = winef, column = c(3:4), "ProductName + (1|CJ)+ (1|ProductName:CJ)")
\dontrun{
export.AOV(res.AOV)
}
mycolor <- c("#EDF8E9", "#C7E9C0", "#A1D99B", "#74C476", "#41AB5D", "#238B45", "#005A32")
graph.AOV(res.AOV, color.graph = mycolor, y.label = "Moyenne")
graph.AOV(res.AOV, title.graph = "Sensory Score", add.moy = TRUE, x.ordered = 1, ymin = 1, ymax = 9)

}
\seealso{
\code{\link[=AOV.Grappe]{AOV.Grappe()}}
}
