% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrata.multidim.R
\name{hrata.multidim}
\alias{hrata.multidim}
\title{hrata.multidim : Performs a multidimensionnal analysis PCA for HRATA and HCATA data}
\usage{
hrata.multidim(
  res.table,
  method = "PCA",
  scale.unit = TRUE,
  niv = 3,
  select = NULL
)
}
\arguments{
\item{res.table}{a dataframe provided by \link{hrata.table}}

\item{method}{"PCA" for HRATA or "CA" for HCATA and HRATA}

\item{scale.unit}{a boolean, if TRUE (value set by default) then data are scaled to unit variance}

\item{niv}{a value to be fixe at 2 if the number of hierarichie is 2}

\item{select}{a value indicating which level are used as active variables  : "cat" for category, "fam" for familly, "att" for terms (by defaul NULL)}
}
\value{
Returns a list similar to \link[FactoMineR:PCA]{FactoMineR::PCA}
}
\description{
Performs a multidimensionnal analysis PCA for HRATA data and HCATA data
}
\examples{
data(rose)
data(rose.attribute)
res.agreg <- hrata.agregation(data = rose, h.table = rose.attribute, crit.agreg = "max")
res.signi <- hrata.signi(res.agreg, seuil = 0.05)
res.table <- hrata.table(res.agreg, type = "dravnieks")
res.PCA <- hrata.multidim(res.table, method = "PCA", scale.unit = FALSE, niv = 3)

}
\seealso{
\code{\link[=hrata.signi]{hrata.signi()}} \code{\link[=hrata.agregation]{hrata.agregation()}} \code{\link[=hrata.table]{hrata.table()}}
}
