% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrata.signi.R
\name{hrata.signi}
\alias{hrata.signi}
\title{hrata.signi : Perform a regression analysis}
\usage{
hrata.signi(res.agreg, select.judge = FALSE, seuil = NULL)
}
\arguments{
\item{res.agreg}{results from \link{hrata.agregation}}

\item{select.judge}{logical. (By default = TRUE) if TRUE subjects who did not consider an attribute were excluded from the analysis of that attribute so that the number of subjects for each attribute varied. If FALSE : all subjects are used fo all attributes}

\item{seuil}{numeric. the significance threshold considered for the analyses of variance. Only attribute under this value will keep in the results (by default 0.05)}
}
\value{
Returns a dataframe with the pvalue for each attributes
}
\description{
Perform a logistic regression was  performed for each attribute, with subject and product as independent factors (Dobson and Barnett, 2008) for HRATA or HCATA data to identify discriminative attibutes
as recommended  in Koenig et al (2023)
}
\examples{
data(rose)
data(rose.attribute)
res.agreg <- hrata.agregation(data = rose, h.table = rose.attribute, crit.agreg = "max")
res.signi <- hrata.signi(res.agreg, seuil = 0.05)
res.table <- hrata.table(res.agreg, type = "dravnieks")
res.PCA <- hrata.multidim(res.table, method = "PCA", scale.unit = FALSE, niv = 3)

}
\seealso{
\code{\link[=AOV.Grappe]{AOV.Grappe()}}
}
