% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrata.table.R
\name{hrata.table}
\alias{hrata.table}
\title{hrata.table : Compute the average table from HRATA result}
\usage{
hrata.table(res.agreg, type = "dravnieks")
}
\arguments{
\item{res.agreg}{results from \link{hrata.agregation}}

\item{type}{aggregation criterion of the average table ("mean" or "dravnieks") by default "dravnieks"}
}
\description{
Compute the dravnieks or mean table from HRATA results
}
\examples{
data(apple)
data(apple.attribute)
res.agreg<-hrata.agregation(data=apple,h.table=apple.attribute,crit.agreg=max)
res.signi<-hrata.signi(res.agreg,seuil=0.05)
res.table<-hrata.table(res.agreg,type="dravnieks")
res.PCA<-hrata.multidim(res.table,method="PCA",scale.unit=FALSE,niv=3)


}
\seealso{
XXXXXX
}
