% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrata.table.R
\name{hrata.table}
\alias{hrata.table}
\title{hrata.table : Compute the average table for HRATA or HCATA multidimensionnal result}
\usage{
hrata.table(res.agreg, select.judge = FALSE, type = "dravnieks")
}
\arguments{
\item{res.agreg}{results from \link{hrata.agregation}}

\item{select.judge}{logical. (By default = TRUE) if TRUE subjects who did not consider an attribute were excluded from the analysis of that attribute so that the number of subjects for each attribute varied. If FALSE : all subjects are used fo all attributes}

\item{type}{aggregation criterion of the average table ("mean" or "dravnieks" or "contingency") by default "dravnieks"}
}
\description{
Compute the dravnieks or mean table from HRATA results or the contingency table for HCATA as defined in Koenig et al (2023)
}
\examples{
data(rose)
data(rose.attribute)
res.agreg<-hrata.agregation(data=rose,h.table=rose.attribute,crit.agreg=max)
res.signi<-hrata.signi(res.agreg,seuil=0.05)
res.table<-hrata.table(res.agreg,type="dravnieks")
res.PCA<-hrata.multidim(res.table,method="PCA",scale.unit=FALSE,niv=3)


}
\seealso{
\code{\link[=hrata.signi]{hrata.signi()}} \code{\link[=hrata.agregation]{hrata.agregation()}} \code{\link[=hrata.multidim]{hrata.multidim()}}
}
