% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PCA.grappe.R
\name{plot.PCA.grappe}
\alias{plot.PCA.grappe}
\title{Draw the Principal Component Analysis (PCA) graphs - Adapted From FactoMineR}
\usage{
\method{plot}{PCA.grappe}(
  x,
  axes = c(1, 2),
  choix = c("ind", "var", "varcor"),
  ellipse = NULL,
  xlim = NULL,
  ylim = NULL,
  habillage = "none",
  col.hab = NULL,
  col.ind = "black",
  col.ind.sup = "blue",
  col.quali = "magenta",
  col.quanti.sup = "blue",
  col.var = "black",
  label = c("all", "none", "ind", "ind.sup", "quali", "var", "quanti.sup"),
  invisible = c("none", "ind", "ind.sup", "quali", "var", "quanti.sup"),
  lim.cos2.var = 0,
  title = NULL,
  palette = NULL,
  autoLab = c("auto", "yes", "no"),
  new.plot = FALSE,
  select = NULL,
  unselect = 0.7,
  shadowtext = FALSE,
  legend = list(bty = "y", x = "topleft"),
  graph.type = c("ggplot", "classic"),
  ggoptions = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class PCA}

\item{axes}{a length 2 vector specifying the components to plot}

\item{choix}{the graph to plot ("ind" for the individuals, "var" for the variables,
"varcor" for a graph with the correlation circle when \code{scale.unit=FALSE})}

\item{ellipse}{boolean (NULL by default), if not null, draw ellipses around the
individuals, and use the results of \code{coord.ellipse}}

\item{xlim}{range for the plotted 'x' values, defaulting to the range of the finite
values of 'x'}

\item{ylim}{range for the plotted 'y' values, defaulting to the range of the finite
values of 'y'}

\item{habillage}{give no color for the individuals ("none"), a color for each
individual ("ind"), or color the individuals among a categorical variable
(give the number of the categorical variable)}

\item{col.hab}{a vector with the color to use for the individuals}

\item{col.ind}{a color for the individuals only if there is not habillage}

\item{col.ind.sup}{a color for the supplementary individuals only if there is not habillage}

\item{col.quali}{a color for the categories of categorical variables only if there
is not habillage}

\item{col.quanti.sup}{a color for the quantitative supplementary variables}

\item{col.var}{a color for the variables}

\item{label}{a list of character for the elements which are labelled (by default,
all the elements are labelled ("ind", "ind.sup", "quali", "var", "quanti.sup"))}

\item{invisible}{string indicating if some points should not be drawn ("ind",
"ind.sup" or "quali" for the individual graph and "var" or "quanti.sup" for
the correlation circle graph)}

\item{lim.cos2.var}{value of the square cosinus under which the variables are not drawn}

\item{title}{string corresponding to the title of the graph (by default NULL and
a title is chosen)}

\item{palette}{the color palette used to draw the points. By default colors are chosen.
If you want to define the colors: \code{palette=palette(c("black","red","blue"))};
or you can use: \code{palette=palette(rainbow(30))}, or in black and white:
\code{palette=palette(gray(seq(0,.9,len=25)))}}

\item{autoLab}{if \code{autoLab="auto"}, equal to "yes" if there are less than 50
elements and "no" otherwise; if "yes", labels are placed in a "good" way
(can be time-consuming if many elements), if "no" elements are placed quickly
but may overlap}

\item{new.plot}{boolean, if TRUE a new graphical device is created; only used
when \code{graph.type="classic"}}

\item{select}{a selection of the elements that are drawn; see the details section}

\item{unselect}{may be either a value between 0 and 1 that gives the transparency
of the unselected objects, or a color (for example \code{unselect="grey60"})}

\item{shadowtext}{boolean; if TRUE put a shadow on the labels; only used when
\code{graph.type="classic"}}

\item{legend}{a list of arguments that defines the legend if needed; see the
arguments of the function \code{legend}}

\item{graph.type}{a character that gives the type of graph: "ggplot" or "classic"}

\item{ggoptions}{a list that gives the graph options when \code{graph.type="ggplot"}
is used. See the options and default values in the details section}

\item{...}{further arguments passed to or from other methods, such as cex, cex.main}
}
\value{
Returns the individuals factor map and the variables factor map.
}
\description{
Plot the graphs for a Principal Component Analysis (PCA) with supplementary individuals,
supplementary quantitative variables and supplementary categorical variables.
}
\details{
The argument \code{autoLab = "yes"} is time-consuming if there are many labels
that overlap. In this case, you can modify the size of the characters in order to
have less overlapping, using for example \code{cex=0.7}.\cr
The \code{select} argument can be used to select a part of the elements that are drawn:\cr
\code{select = 1:5} : elements 1 to 5 are drawn.\cr
\code{select = c("name1","name5")} : elements with names name1 and name5 are drawn.\cr
\code{select = "coord 10"} : the 10 elements with the highest coordinates on the 2 chosen dimensions.\cr
\code{select = "contrib 10"} : the 10 elements with the highest contribution on the 2 dimensions.\cr
\code{select = "cos2 5"} : the 5 elements with the highest cos2 on the 2 dimensions.\cr
\code{select = "dist 8"} : the 8 elements with the highest distance to the center of gravity.\cr
\cr
\code{ggoptions} default values:\cr
size = 4, point.shape = 19, line.lty = 2, line.lwd = 0, line.color = "black",\cr
segment.lty = 1, segment.lwd = 0, circle.lty = 1, circle.lwd = 0,\cr
circle.color = "black", low.col.quanti = "blue", high.col.quanti = "red3"
}
\examples{
data("decathlon", package = "FactoMineR")
res.pca <- FactoMineR::PCA(decathlon, quanti.sup = 11:12, quali.sup = 13)
plot(res.pca, habillage = 13, cex = 0.8)

}
\seealso{
\code{\link[FactoMineR]{PCA}}
}
\author{
Francois Husson \email{francois.husson@institut-agro.fr}
}
