% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.get.binary.R
\name{transfert.MV.BM}
\alias{transfert.MV.BM}
\title{Decomposition d'un descripteur hierarchique (format ConsoTextPlorer)}
\usage{
transfert.MV.BM(x, tres = FALSE)
}
\arguments{
\item{x}{Caractere. Chaine a decomposer, au format
\code{"dimension:concept/descripteur"}. Peut etre \code{NA}.}

\item{tres}{Logique. Si \code{FALSE} (defaut), le suffixe \verb{_tres} est
supprime des composantes \code{concept} et \code{descripteur}.
Si \code{TRUE}, il est conserve.}
}
\value{
Une matrice de dimensions 1 x 3 contenant, dans l'ordre :
\code{dimension}, \code{concept}, \code{descripteur}. Si \code{x} est \code{NA}, retourne
la matrice \code{c("rien", "rien", "rien")}.
}
\description{
Decompose une chaine de caracteres au format ConsoTextPlorer
(\code{"dimension:concept/descripteur"}) en ses trois composantes :
dimension, concept et descripteur.
}
\details{
Le format attendu est \code{"dimension:concept/descripteur"}, ou :
\itemize{
\item \code{":"} separe la dimension du concept ;
\item \code{"/"} separe le concept du descripteur.
}
Si le descripteur vaut \code{"_"}, il est remplace par la valeur du concept.
}
\examples{
transfert.MV.BM("texture:croustillant/tres_croustillant", tres = FALSE)
#>      [,1]      [,2]           [,3]
#> [1,] "texture" "croustillant" "croustillant"

transfert.MV.BM("texture:croustillant/tres_croustillant", tres = TRUE)
#>      [,1]      [,2]           [,3]
#> [1,] "texture" "croustillant" "tres_croustillant"

transfert.MV.BM(NA)
#>      [,1]    [,2]    [,3]
#> [1,] "rien"  "rien"  "rien"

}
\seealso{
\code{\link{get.binary}}, \code{\link{data.preprocess.fc}}
}
