% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Var.Grappe.R
\name{Var.Grappe}
\alias{Var.Grappe}
\title{Var.Grappe : Checking and encoding dataset}
\usage{
Var.Grappe(x, column = 0, type = NULL, filter = NULL, verbose = TRUE)
}
\arguments{
\item{x}{dataframe}

\item{column}{the numbers or the name of the columns to change the type}

\item{type}{expected type of the columns to be changed : "factor", "numeric" or "integer". NULL by default}

\item{filter}{character or numeric corresponding to searching name or number of column. Apply VarGrappe to a set of a fltered
columns but x is not filtered in the thrown object}

\item{verbose}{logical display of the results. TRUE by default}
}
\description{
Check dataset and  encode easily the vector (variables) to factor, integer or numeric
}
\examples{
data(wine)
Var.Grappe(wine)
Var.Grappe(wine, filter = "O_")
winef <- Var.Grappe(wine, column = c(1:2), type = "factor")

}
